requirements.txt
Django==1.7.1
argparse==1.2.1
django-colorful==1.0.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
django-bootstrap-form==3.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
